Ai et al., Cardiomyocytes Single Cell Proteomics, MCP (2025)

The manuscript analysed cardiomyocytes derived from iPSC cells and adult hearts. This annotated QFeatures objects contains the complete data, as analysed with the scp package. The re-analysis was performed as part of the EuBIC 2025 developer meeting hackathon dedicated to single-cell proteomics data (https://lgatto.github.io/2025-EuBIC-SCP-hackathon/).

The data contains a total of 304 assay. The 299 first assays contain the individual single cardiomyocytes acquistions. The following assays contain the joined precursor data, the log-transformed precuror data, the peptide data, the protein data, and the scplainer-modelled precursor data (see the link above for details).

The QFeatures object was generated from the DIA-NN report file, available from the MsDataHub package (see ?MsDataHub::Ai2025_aCMs_report.tsv). The report file is downloaded from the MassIVE MSV000094438 (doi:10.25345/C5T727S7Q) repository.

Usage

ai2025a

Format

An object of class function of length 1.

References

Ai, Lizhuo, Aleksandra Binek, Vladimir Zhemkov, Jae Hyung Cho, Ali Haghani, Simion Kreimer, Edo Israely, et al. 2025. “Single Cell Proteomics Reveals Specific Cellular Subtypes in Cardiomyocytes Derived from Human iPSCs and Adult Hearts.” Molecular & Cellular Proteomics: MCP, no. 100910 (January): 100910.

Examples

# \donttest{
ai2025a()
#> see ?scpdata and browseVignettes('scpdata') for documentation
#> loading from cache
#> An instance of class QFeatures (type: scp) with 304 sets:
#> 
#>  [1] 221019_CM_SUBJECT-03_CMs_Subject3_Lvendo_A13_BA14_1_7445.d: SingleCellExperiment with 6082 rows and 1 columns 
#>  [2] 221019_CM_SUBJECT-03_CMs_Subject3_Lvendo_A14_BA15_1_7446.d: SingleCellExperiment with 3958 rows and 1 columns 
#>  [3] 221019_CM_SUBJECT-03_CMs_Subject3_Lvendo_A15_BA16_1_7447.d: SingleCellExperiment with 5729 rows and 1 columns 
#>  ...
#>  [302] peptides: SingleCellExperiment with 10277 rows and 299 columns 
#>  [303] proteins: SingleCellExperiment with 2044 rows and 299 columns 
#>  [304] precursors_stats: SingleCellExperiment with 11731 rows and 278 columns 
# }