Hu et al, 2023 (The Journal of Physical Chemistry B): Correlated protein modules

They demonstrate the correlations between the levels of pairs of proteins in single-cell proteomics (SCP) at steady state. In measuring pairwise correlations among 1000 proteins in a population of K562 cells and oocytes, they observed many correlated protein modules (CPMs) that are functionally involved in certain biological functions. Certain CPMs are specific to a particular cell type, some common to different cell types. Additionally, compared to single-cell transcriptomics and bulk proteomics, protein correlations are functionally and experimentally more significant in SCP than those corresponding mRNAs.

Usage

hu2023_K562

Format

Two SingleCellExperiment::SingleCellExperiment objects:

proteins_K562: protein data containing quantitative data for 1249 proteins and 69 single-cells with zero imputation.

proteins_oocyte: protein data containing quantitative data for 3422 proteins and 137 single-cells with zero imputation.

The colData(hu2023_oocyte()) contains cell type annotation. The colData(hu2023_K562()) contains cell type annotation.

Source

The oocyte data were downloaded from the Shared File The K563 cells protein data downloaded from the GitHub (https://github.com/dionezhang/CPM/blob/master/ProteinAbundance.Rdata) The raw data and the quantification data can also be found in the MassIVE repository MSV000089625: ftp://MSV000089625@massive.ucsd.edu/.

Acquisition protocol

The data were acquired using the following setup. More information can be found in the source article (see References).

Cell isolation: K562 cells were re-suspended and washed in cold PBS. Single cells/10 cells were sorted into 96-well plates using a FACSAria instrument. Oocyte-cumulus complexes from C57/6J mice were collected after PMSG and HCG injections, with hyaluronidase used to remove cumulus cells. All samples stored at -80 degrees Celsius.

Sample preparation Cells were digested with trypsin at 37 degrees Celsius for 3 hours. For label-free proteomics, digestion was terminated by adding 0.43% TFA and 1% ACN in water, followed by drying in a concentrator. Peptides were resuspended in 0.1% TFA and 1% ACN, and then transferred to sample tubes for LC-MS/MS analysis.

Separation: 4 microliters of peptide digests were injected into a high-performance chromatography column (IonOpticks) and separated at a flow rate of 100 nL/min using a nanoflow liquid chromatography system. The effective gradient was 70 mins, allowing 16 cells per day.

Ionization: Peptides were analyzed using an Orbitrap Eclipse mass spectrometer with a FAIMS Pro interface. FAIMS compensation voltages of -55 and -70 V were applied, with a 1-second cycle time for both voltages.

Mass spectrometry: MS spectra were acquired with the Orbitrap analyzer, while MS/MS spectra were acquired with a linear ion trap analyzer. The maximum ion injection time for MS/MS was 200 ms.

Data analysis: MS raw files were searched against the UniProt human protein database and an in-house contamination database using Proteome Discoverer(2.4). Label-free quantification was based on peak intensity with the match-between-runs (MBR) feature enabled.

Data collection

The oocyte protein data shared by the author and it is accessible from the Shared File The K563 protein data is accessible from the GitHub (https://github.com/dionezhang/CPM/blob/master/ProteinAbundance.Rdata).

DataMatrix-oocyte-20240614.csv: normalized imputed protein matrix

ProteinAbundance.Rdata: protein matrices (normalized, log transformed)

We initialized an empty QFeatures object and added the corresponding protein assays as SingleCellExperiment::SingleCellExperiment objects.

The oocyte protein data were exported from the shared link as (DataMatrix-oocyte-20240614.csv). The data were formatted to a SingleCellExperiment::SingleCellExperiment object and the SampleType information were added as only metadata, and stored in the colData. The object is then added to the QFeatures::QFeatures object.

The 562 cells protein data were downloaded from the GitHub link and loaded to the memory. The Norm object were formatted to a SingleCellExperiment::SingleCellExperiment object and the SampleType information were added as only metadata, and stored in the colData. The object is then added to the QFeatures::QFeatures object.

References

Hu, M., Zhang, Y., Yuan, Y., Ma, W., Zheng, Y., Gu, Q., & Xie, X. S. 2023. “Correlated protein modules revealing functional coordination of interacting proteins are detected by single-cell proteomics.”. The Journal of Physical Chemistry B, (link to article).

Examples

# \donttest{
hu2023_oocyte()
#> see ?scpdata and browseVignettes('scpdata') for documentation
#> loading from cache
#> require(“SingleCellExperiment”)
#> class: SingleCellExperiment 
#> dim: 3422 137 
#> metadata(0):
#> assays(1): ''
#> rownames(3422): P16125 Q8K3V4 ... Q91VU8 Q08274
#> rowData names(0):
#> colnames(137): h1 h3 ... z83 z86
#> colData names(1): SampleType
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
hu2023_K562()
#> see ?scpdata and browseVignettes('scpdata') for documentation
#> loading from cache
#> class: SingleCellExperiment 
#> dim: 1249 69 
#> metadata(0):
#> assays(1): ''
#> rownames(1249): P49327 P08238 ... Q969Q0 O15400
#> rowData names(2): Description protein
#> colnames(69): F1 F2 ... F94 F96
#> colData names(1): SampleType
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
# }