Wu et al., 2026 (Nature Biotechnology): single-cell proteomic landscape of the developing human brain

Label-free single-cell proteomics data from prenatal human brain tissue (gestational weeks 13, 15 and 19), generated to characterize cell-type proteomes and developmental trajectories at single-cell resolution.

Usage

wu2026

Format

A QFeatures::QFeatures object with 2313 sets, each set being a SingleCellExperiment::SingleCellExperiment object:

Set 1-2310: log-transformed precursor-level single-cell data (one column per set) from DIA-NN reports.

peptides: peptide-level data containing quantitative data for 17755 peptides in 2310 single cells.

imported_proteins: protein-level table imported from the authors' protein report, containing 4401 proteins in 2310 single cells.

proteins: protein-level data containing quantitative data for 5107 proteins in 2310 single cells.

Sample annotation is stored in colData(wu2026()).

Source

The data were downloaded from the PRIDE repository with accession ID PXD071075: https://www.ebi.ac.uk/pride/archive/projects/PXD071075

Acquisition protocol

The data were acquired using the following setup. More information can be found in the source article (see References).

Sample isolation: fresh prenatal human brain tissue was dissociated and single cells were sorted by FACS (Sony SH800) into 384-well low-bind plates.

Sample preparation: cells were lysed in low-volume TEAB-based buffer, digested overnight with trypsin, acidified with trifluoroacetic acid, and desalted on EvoTip Pure.

Liquid chromatography: peptides were separated on an IonOpticks Aurora Elite XT C18 nano-LC column (15 cm x 75 micrometer) on an Evosep One using a 31-minute gradient.

Mass spectrometry: Orbitrap Eclipse Tribrid with FAIMS Pro was used in DIA mode.

Raw data processing: DIA-NN against the UniProt human reference proteome.

Data collection

The QFeatures object was built from the DIA-NN peptide and protein report tables and a sample annotation table. PSM-level data were imported with scp::readSCP(), log-transformed, aggregated to peptide-level features, and joined across single-cell runs. The protein report table was imported as a SingleCellExperiment::SingleCellExperiment, added as an set, and linked to peptide-level data.

The script to reproduce the QFeatures object is available at inst/scripts/make-data_wu2026.R.

References

Wu, T., Jiang, L., Mukhtar, T., Wang, L., Jian, R., Wang, C., Trinh, T., Kriegstein, A. R., Snyder, M., and Li, J. 2026. "Single-cell proteomic landscape of the developing human brain." Nature Biotechnology. Link to article

Examples

# \donttest{
wu2026()
#> see ?scpdata and browseVignettes('scpdata') for documentation
#> loading from cache
#> An instance of class QFeatures (type: scp) with 2313 sets:
#> 
#>  [1] S1_10: SummarizedExperiment with 239 rows and 1 columns 
#>  [2] S1_100: SummarizedExperiment with 849 rows and 1 columns 
#>  [3] S1_1000: SummarizedExperiment with 239 rows and 1 columns 
#>  ...
#>  [2311] peptides: SummarizedExperiment with 17755 rows and 2310 columns 
#>  [2312] imported_proteins: SummarizedExperiment with 4401 rows and 2310 columns 
#>  [2313] proteins: SummarizedExperiment with 5107 rows and 2310 columns 
# }