Dou et al. 2019 (Anal. Chem.): HeLa lysates

Single-cell proteomics using nanoPOTS combined with TMT multiplexing. It contains quantitative information at PSM and protein level. The samples are commercial Hela lysates diluted to single-cell amounts (0.2 ng). The boosting wells contain the same digest but at higher amount (10 ng).

dou2019_lysates

Format

A QFeatures object with 3 assays, each assay being a SingleCellExperiment object:

Hela_run_1: PSM data with 10 columns corresponding to the TMT-10plex channels. Columns hold quantitative information for HeLa lysate samples (either 0, 0.2 or 10ng). This is the data for run 1.

Hela_run_1: PSM data with 10 columns corresponding to the TMT-10plex channels. Columns hold quantitative information for HeLa lysate samples (either 0, 0.2 or 10ng). This is the data for run 2.

proteins: protein data containing quantitative data for 1641 proteins and 20 samples (run 1 and run 2 combined).

Sample annotation is stored in colData(dou2019_lysates()). The description of the rowData fields for the PSM data can be found in the MaxQuant documentation.

Source

The PSM data can be downloaded from the massIVE repository MSV000084110. FTP link: ftp://massive.ucsd.edu/MSV000084110/

The protein data can be downloaded from the ACS Publications website (Supporting information section).

Acquisition protocol

The data were acquired using the following setup. More information can be found in the source article (see References).

Cell isolation: commercially available HeLa protein digest (Thermo Scientific).

Sample preparation performed using the nanoPOTs device. Protein extraction (DMM + TCEAP) + alkylation (IAA) + Lys-C digestion + trypsin digestion + TMT-10plex labeling and pooling.

Separation: nanoLC (Dionex UltiMate with an in-house packed 50cm x 30um LC columns; 50nL/min)

Ionization: ESI (2,000V)

Mass spectrometry: Thermo Fisher Orbitrap Fusion Lumos Tribrid (MS1 accumulation time = 50ms; MS1 resolution = 120,000; MS1 AGC = 1E6; MS2 accumulation time = 246ms; MS2 resolution = 60,000; MS2 AGC = 1E5)

Data analysis: MS-GF+ + MASIC (v3.0.7111) + RomicsProcessor (custom R package)

Data collection

The PSM data were collected from the MassIVE repository MSV000084110 (see Source section). The downloaded files are:

Hela_run_*_msgfplus.mzid: the MS-GF+ identification result files

Hela_run_*_ReporterIons.txt: the MASIC quantification result files

For each batch, the quantification and identification data were combined based on the scan number (common to both data sets). The combined datasets for the different runs were then concatenated feature-wise. The sample annotation table was manually created based on the available information provided in the article. The data were then converted to a QFeatures object using the scp::readSCP function.

The protein data were downloaded from Supporting information section from the publisher's website (see Sources). The data is supplied as an Excel file ac9b03349_si_003.xlsx. The file contains 7 sheets from which we only took the sheet 6 (named 5 - Run 1 and 2 raw data) with the combined protein data for the two runs. We converted the data to a SingleCellExperiment object and added the object as a new assay in the QFeatures dataset (containing the PSM data). Links between the proteins and the corresponding PSM were created.

References

Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample Preparation Platform.” Analytical Chemistry, September (link to article).

See also

Examples

dou2019_lysates()

#> see ?scpdata and browseVignettes('scpdata') for documentation

#> loading from cache

#> An instance of class QFeatures containing 3 assays:
#>  [1] Hela_run_1: SingleCellExperiment with 35224 rows and 10 columns 
#>  [2] Hela_run_2: SingleCellExperiment with 35080 rows and 10 columns 
#>  [3] proteins: SingleCellExperiment with 1641 rows and 20 columns