Single-Cell Proteomics Data Package

The scpdata package distributes mass spectrometry-based single-cell proteomics datasets. The datasets were collected from published work and formatted to a standardized data framework. The scp frameworks stores the expression data for different MS levels (identified spectrum, peptide, or protein) in separate assays. Each assay is an object of class SingleCellExperiment that allows easy integration with state-of-the-art single-cell analysis tools. All assays are contained in a single object of class QFeatures. An overview of the data structure is shown provided in the scp package.

The scpdata() function returns a summary table with all currently available datasets in the package. More information about the data content and the data collection can be found in the corresponding manual pages.

scpdata()

Value

A DataFrame table containing a summary of the available datasets.

See also

Author

Christophe Vanderaa

Examples

## List available datasets and their metadata 
scpdata()

#> snapshotDate(): 2021-01-28

#> DataFrame with 11 rows and 15 columns
#>                title  dataprovider       species taxonomyid      genome
#>          <character>   <character>   <character>  <integer> <character>
#> EH3899 specht2019... SlavovLab ... Homo sapie...       9606          NA
#> EH3900 specht2019... SlavovLab ... Homo sapie...       9606          NA
#> EH3901 dou2019_ly...       MassIVE Homo sapie...       9606          NA
#> EH3902 dou2019_mo...       MassIVE Mus muscul...      10090          NA
#> EH3903 dou2019_bo...       MassIVE Mus muscul...      10090          NA
#> EH3904    zhu2018MCP         PRIDE Rattus nor...      10116          NA
#> EH3905 zhu2018NC_...         PRIDE Homo sapie...       9606          NA
#> EH3906 zhu2018NC_...         PRIDE Homo sapie...       9606          NA
#> EH3907 zhu2018NC_...         PRIDE Homo sapie...       9606          NA
#> EH3908    cong2020AC         PRIDE Homo sapie...       9606          NA
#> EH3909     zhu2019EL         PRIDE Gallus gal...       9031          NA
#>          description coordinate_1_based    maintainer rdatadateadded
#>          <character>          <integer>   <character>    <character>
#> EH3899 SCP expres...                  1 Christophe...     2020-11-05
#> EH3900 SCP expres...                  1 Christophe...     2020-11-05
#> EH3901 SCP expres...                  1 Christophe...     2020-11-05
#> EH3902 SCP expres...                  1 Christophe...     2020-11-05
#> EH3903 SCP expres...                  1 Christophe...     2020-11-05
#> EH3904 Near SCP e...                  1 Christophe...     2020-11-05
#> EH3905 Near SCP e...                  1 Christophe...     2020-11-05
#> EH3906 Near SCP e...                  1 Christophe...     2020-11-05
#> EH3907 Near SCP e...                  1 Christophe...     2020-11-05
#> EH3908 SCP expres...                  1 Christophe...     2020-11-05
#> EH3909 SCP expres...                  1 Christophe...     2020-11-05
#>        preparerclass                                          tags  rdataclass
#>          <character>                                        <list> <character>
#> EH3899       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3900       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3901       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3902       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3903       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3904       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3905       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3906       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3907       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3908       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3909       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#>            rdatapath     sourceurl  sourcetype
#>          <character>   <character> <character>
#> EH3899 scpdata/sp... https://sc...         CSV
#> EH3900 scpdata/sp... https://sc...         CSV
#> EH3901 scpdata/do... ftp://mass...    XLS/XLSX
#> EH3902 scpdata/do... ftp://mass...    XLS/XLSX
#> EH3903 scpdata/do... ftp://mass...    XLS/XLSX
#> EH3904 scpdata/zh... ftp://ftp....         TXT
#> EH3905 scpdata/zh... ftp://ftp....         TXT
#> EH3906 scpdata/zh... ftp://ftp....         TXT
#> EH3907 scpdata/zh... ftp://ftp....         TXT
#> EH3908 scpdata/co... ftp://ftp....         TXT
#> EH3909 scpdata/zh... ftp://ftp....         TXT

## Load data using the ExperimentHub interface
hub <- ExperimentHub()

#> snapshotDate(): 2021-01-28

## Download the data set of interest using ExperimentHub indexing
hub[["EH3899"]]

#> see ?scpdata and browseVignettes('scpdata') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> An instance of class QFeatures containing 179 assays:
#>  [1] 190222S_LCA9_X_FP94AA: SingleCellExperiment with 2823 rows and 11 columns 
#>  [2] 190222S_LCA9_X_FP94AB: SingleCellExperiment with 4297 rows and 11 columns 
#>  [3] 190222S_LCA9_X_FP94AC: SingleCellExperiment with 4956 rows and 11 columns 
#>  ...
#>  [177] 191110S_LCB7_X_APNOV16plex2_Set_9: SingleCellExperiment with 4626 rows and 16 columns 
#>  [178] peptides: SingleCellExperiment with 9208 rows and 1018 columns 
#>  [179] proteins: SingleCellExperiment with 2772 rows and 1018 columns 

## Download the same data set using the build-in function
specht2019v2()

#> snapshotDate(): 2021-01-28

#> see ?scpdata and browseVignettes('scpdata') for documentation

#> loading from cache

#> An instance of class QFeatures containing 179 assays:
#>  [1] 190222S_LCA9_X_FP94AA: SingleCellExperiment with 2823 rows and 11 columns 
#>  [2] 190222S_LCA9_X_FP94AB: SingleCellExperiment with 4297 rows and 11 columns 
#>  [3] 190222S_LCA9_X_FP94AC: SingleCellExperiment with 4956 rows and 11 columns 
#>  ...
#>  [177] 191110S_LCB7_X_APNOV16plex2_Set_9: SingleCellExperiment with 4626 rows and 16 columns 
#>  [178] peptides: SingleCellExperiment with 9208 rows and 1018 columns 
#>  [179] proteins: SingleCellExperiment with 2772 rows and 1018 columns