Single-Cell Proteomics Data Package

The scpdata package distributes mass spectrometry-based single-cell proteomics datasets. The datasets were collected from published work and formatted to a standardized data framework. The scp frameworks stores the expression data for different MS levels (identified spectrum, peptide, or protein) in separate assays. Each assay is an object of class SingleCellExperiment that allows easy integration with state-of-the-art single-cell analysis tools. All assays are contained in a single object of class QFeatures. An overview of the data structure is shown provided in the scp package.

The scpdata() function returns a summary table with all currently available datasets in the package. More information about the data content and the data collection can be found in the corresponding manual pages.

scpdata()

Value

A DataFrame table containing a summary of the available datasets.

See also

Author

Christophe Vanderaa

Examples

## List available datasets and their metadata 
scpdata()
#> DataFrame with 26 rows and 15 columns
#>                title  dataprovider       species taxonomyid      genome
#>          <character>   <character>   <character>  <integer> <character>
#> EH3899 specht2019... SlavovLab ... Homo sapie...       9606          NA
#> EH3900 specht2019... SlavovLab ... Homo sapie...       9606          NA
#> EH3901 dou2019_ly...       MassIVE Homo sapie...       9606          NA
#> EH3902 dou2019_mo...       MassIVE Mus muscul...      10090          NA
#> EH3903 dou2019_bo...       MassIVE Mus muscul...      10090          NA
#> ...              ...           ...           ...        ...         ...
#> EH9450 gregoire20...         PRIDE Homo sapie...       9606          NA
#> EH9477      khan2023       MassIVE Homo sapie...       9606          NA
#> EH9487     guise2024       MassIVE Homo sapie...       9606          NA
#> EH9497 petrosius2...     Dataverse Homo sapie...       9606          NA
#> EH9498 petrosius2...     Dataverse Homo sapie...       9606          NA
#>          description coordinate_1_based    maintainer rdatadateadded
#>          <character>          <integer>   <character>    <character>
#> EH3899 SCP expres...                  1 Christophe...     2020-11-05
#> EH3900 SCP expres...                  1 Christophe...     2020-11-05
#> EH3901 SCP expres...                  1 Christophe...     2020-11-05
#> EH3902 SCP expres...                  1 Christophe...     2020-11-05
#> EH3903 SCP expres...                  1 Christophe...     2020-11-05
#> ...              ...                ...           ...            ...
#> EH9450 Single-cel...                  1 Samuel Gre...     2024-02-12
#> EH9477 Single-cel...                  1 Enes Sefa ...     2024-03-08
#> EH9487 Single-cel...                  1 Christophe...     2024-04-08
#> EH9497 Mouse embr...                  1 Enes Sefa ...     2024-04-11
#> EH9498 Single-cel...                  1 Samuel Gre...     2024-04-11
#>        preparerclass                                          tags  rdataclass
#>          <character>                                        <AsIs> <character>
#> EH3899       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3900       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3901       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3902       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> EH3903       scpdata Experiment...,Expression...,Experiment...,...   QFeatures
#> ...              ...                                           ...         ...
#> EH9450       scpdata      Expression...,MassSpectr...,Proteome,...   QFeatures
#> EH9477       scpdata      Expression...,MassSpectr...,Proteome,...   QFeatures
#> EH9487       scpdata      Expression...,MassSpectr...,Proteome,...   QFeatures
#> EH9497       scpdata      Expression...,MassSpectr...,Proteome,...   QFeatures
#> EH9498       scpdata      Expression...,MassSpectr...,Proteome,...   QFeatures
#>            rdatapath     sourceurl  sourcetype
#>          <character>   <character> <character>
#> EH3899 scpdata/sp... https://sc...         CSV
#> EH3900 scpdata/sp... https://sc...         CSV
#> EH3901 scpdata/do... ftp://mass...    XLS/XLSX
#> EH3902 scpdata/do... ftp://mass...    XLS/XLSX
#> EH3903 scpdata/do... ftp://mass...    XLS/XLSX
#> ...              ...           ...         ...
#> EH9450 scpdata/gr... https://ww...         TXT
#> EH9477 scpdata/kh... https://dr...         TXT
#> EH9487 scpdata/gu... ftp://mass...         TXT
#> EH9497 scpdata/pe... https://da...         TXT
#> EH9498 scpdata/pe... https://da...         TXT

## Load data using the ExperimentHub interface
hub <- ExperimentHub()

if (FALSE) {
## Download the data set of interest using ExperimentHub indexing
hub[["EH7711"]]
## Download the same data set using the build-in function
leduc2022()
}