Zhu et al. 2018 (Nat. Comm.): HeLa titration

Near single-cell proteomics data of HeLa samples containing different number of cells. There are three groups of cell concentrations: low (10-14 cells), medium (35-45 cells) and high (137-141 cells). The data also contain measures for blanks, HeLa lysates (50 cell equivalent) and 2 cancer cell line lysates (MCF7 and THP1, 50 cell equivalent).

zhu2018NC_hela

Format

A QFeatures object with 2 assays:

peptides: a SingleCellExperiment object containing the quantitative data for 37795 peptides in 21 samples.

proteins: a SingleCellExperiment object containing the quantitative (summed intensities) data for 3984 proteins in 21 samples. Note that the iBAQ and LFQ intensities computed by MaxQuant are available from the rowData.

Sample annotation is stored in colData(zhu2018NC_hela()).

Source

The PSM data can be downloaded from the PRIDE repository PXD006847. FTP link: ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2018/01/PXD006847

Acquisition protocol

The data were acquired using the following setup. More information can be found in the original article (see References).

Cell isolation: HeLa cell concentration was adjusted by serial dilution and cell counting was performed manually using an inverted microscope.

Sample preparation performed using the nanoPOTs device. Protein extraction using RapiGest (+ DTT) + alkylation (IAA) + Lys-C digestion + cleave RapiGest (formic acid).

Separation: nanoACQUITY UPLC pump (60nL/min) with an Self-Pack PicoFrit 70cm x 30um LC columns.

Ionization: ESI (1,900V).

Mass spectrometry: Thermo Fisher Orbitrap Fusion Lumos Tribrid. MS1 settings: accumulation time = 246ms; resolution = 120,000; AGC = 1E6. MS/MS settings, depend on the sample size, excepted for the AGC = 1E5. Blank and approx. 10 cells (time = 502ms; resolution = 240,000), approx. 40 cells (time = 246ms; resolution = 120,000), approx. 140 cells (time = 118ms; resolution = 60,000).

Data analysis: MaxQuant (v1.5.3.30) + Perseus + OriginLab 2017

Data collection

The PSM data were collected from the PRIDE repository (accession ID: PXD006847). We downloaded the CulturedCells_peptides.txt file containing the combined identification and quantification results. The sample annotation was inferred from the names of columns holding the quantification data and the information in the article. The data were then converted to a QFeatures object using the scp::readSCP function.

References

Zhu, Ying, Paul D. Piehowski, Rui Zhao, Jing Chen, Yufeng Shen, Ronald J. Moore, Anil K. Shukla, et al. 2018. “Nanodroplet Processing Platform for Deep and Quantitative Proteome Profiling of 10-100 Mammalian Cells.” Nature Communications 9 (1): 882 (link to article).

See also

Examples

zhu2018NC_hela()
#> see ?scpdata and browseVignettes('scpdata') for documentation
#> loading from cache
#> An instance of class QFeatures containing 2 assays:
#>  [1] peptides: SingleCellExperiment with 37795 rows and 21 columns 
#>  [2] proteins: SingleCellExperiment with 3984 rows and 21 columns