Zhu et al. 2018 (Nat. Comm.): HeLa lysates

Near single-cell proteomics data of HeLa lysates at different concentrations (10, 40 and 140 cell equivalent). Each concentration is acquired in triplicate.

zhu2018NC_lysates

Format

A QFeatures object with 1 assay called peptides. It is a SingleCellExperiment object containing the quantitative data for 14921 peptides in 9 samples.

Sample annotation is stored in colData(zhu2018NC_lysates()).

Source

The PSM data can be downloaded from the PRIDE repository PXD006847. The source link is: ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2018/01/PXD006847

Acquisition protocol

The data were acquired using the following setup. More information can be found in the original article (see References).

Cell isolation: HeLas were collected from cell cultures.

Sample preparation performed in bulk (5E5 cells/mL). Protein extraction using RapiGest (+ DTT) + dilution to target concentration + alkylation (IAA) + Lys-C digestion + trypsin digestion + cleave RapiGest (formic acid).

Separation: nanoACQUITY UPLC pump (60nL/min) with an Self-Pack PicoFrit 70cm x 30um LC columns.

Ionization: ESI (1,900V).

Mass spectrometry: Thermo Fisher Orbitrap Fusion Lumos Tribrid. MS1 settings: accumulation time = 246ms; resolution = 120,000; AGC = 1E6. MS/MS settings, depend on the sample size, excepted for the AGC = 1E5. Blank and approx. 10 cells (time = 502ms; resolution = 240,000), approx. 40 cells (time = 246ms; resolution = 120,000), approx. 140 cells (time = 118ms; resolution = 60,000).

Data analysis: MaxQuant (v1.5.3.30) + Perseus + OriginLab 2017.

Data collection

The PSM data were collected from the PRIDE repository (accession ID: PXD006847). We downloaded the CulturedCells_peptides.txt file containing the combined identification and quantification results. The sample annotation was inferred from the names of columns holding the quantification data and the information in the article. The data were then converted to a QFeatures object using the scp::readSCP function.

References

Zhu, Ying, Paul D. Piehowski, Rui Zhao, Jing Chen, Yufeng Shen, Ronald J. Moore, Anil K. Shukla, et al. 2018. “Nanodroplet Processing Platform for Deep and Quantitative Proteome Profiling of 10-100 Mammalian Cells.” Nature Communications 9 (1): 882 (link to article).

See also

Examples

zhu2018NC_lysates()

#> snapshotDate(): 2021-01-28

#> see ?scpdata and browseVignettes('scpdata') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> An instance of class QFeatures containing 1 assays:
#>  [1] peptides: SingleCellExperiment with 14921 rows and 9 columns